+++ bobweb/bob/test/daily_question/test_dq_command_activities.py
@@ -1,6 +1,5 @@
 import datetime
 import os
-from bobweb.bob import main
 import django
 from django.test import TestCase
 from telegram import ReplyMarkup
